requirements.txt
MetaTrader5==5.0.43
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
schedule>=1.1.0
flask>=2.0.0
requests>=2.25.0